Add repodata mirror type
This change adds a 'type' parameter for mirror to support
transparent repodata based crawling. When set to 'repodata'
the tool fetchs repomd.xml and primary.xml.gz file to list
all the required files for a functional rpm mirror.

To make it explicit, the legacy get_uri_list is renamed to
get_weblisting_uri_list.